Update to reflect the 2020 course changes
This module doesn't exist anymore, and a new module called 'Electronics and Information Technology' has taken its place and course code. This module is taught by Éamonn Ó Nualáinn and contains topics from both Electrotechnology and Telecommunications I.  This module was first taught in 2019-2020 under CSU11031-2, and was taught in subsequent years as just CSU11031. Based on my limited understanding of what the course was like in 2019 (I'm a first year), I believe that this module was implemented to replace the previous two, and so should take their place on the website.

This PR is my best attempt to account for this - I've linked the (single) available paper and moved the past papers from Telecommunications I and Electrotechnology onto this page in an 'Archive' section, as they may be useful for students.

If this PR is pushed, the Electrotechnology page should be deleted as it offers no utility and is for a defunct module.

I will ask Eamonn on Monday for the past papers from 2022 and 2021 and see if I can add them to the site.

I don't fully understand the formatting behind this site so if this isn't set up properly then deny this PR and tell me what I have to do.