Adding link to my Functional Programming repo
These solutions are from the past 10 or so years, down to 2012 from 2022, the solutions are split into topics. These solutions were used by many of my peers before the exam, and there were some smaller fixes done. Most of the solutions should be correct and most of my classmates said it helped greatly. These solutions differ to nating's solutions, as it attempts to bring in context in some problems for how I found the solution. It also has pdf's/org mode viewable files, which are both supported by github and should give a better experience rather than a single .hs file. In "Monad" questions, I give some code for how to run the code, and what assumptions are being made. There is a section of solutions "Polymorphism" which is a few examples of code which have not been on the exam until the year we sat, so I did up some examples before the exam. My solutions do not contain any copyrighted material and is solely my work. The first two questions of most topics have the same year, this is because the lecturer said it was from a certain year and it did not match the year for on cs-exams.