Adding a solutions to SSE with a link to my year 3 solutions github
My link has 13 years of SSE solutions.

GitHub actions uses the code to check that the commit ensures 
compilation of the files. 

The code is testable by compiling a make file and running the main file. Each year the main file will compare against vectorised and non vectorised code with the difference for comparison.