Updated formal-verification.md
Added a resource with formal verification solutions. This solution is a repository I created in year 4 semester 1, which has solutions to 3 topics that come up in the exam. As the README states, it contains 8/9 years of solutions for those topics, with pdf files generated from the .tex files for reading easily.

Update formal-verification.md